% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geniusR.R
\name{add_genius}
\alias{add_genius}
\title{Add lyrics to a data frame}
\usage{
add_genius(data, artist, type_group, type = "album")
}
\arguments{
\item{data}{This is a dataframe with one column for the artist name, and the other column being either the track title or the album title.}

\item{artist}{This is the column which has artist title information}

\item{type_group}{This is the column that has either album titles or track titles.}

\item{type}{This is a single value character string of either `"album"` or `"lyrics"`. This tells the function what kind of lyrics to pull. This needs to be in line with `type_group`}
}
\description{
This function is to be used to build on a data frame with artist and album/track information. Ideal via the spotifyr package.
}
\examples{
\dontrun{
### Example with 2 different artists and albums
library(dplyr)
library(tibble)

artist_albums <- tribble(
 ~artist, ~album,
 "J. Cole", "KOD",
 "Sampha", "Process"
)


artist_albums \%>\%
 add_genius(artist, album)


## Example with 2 different artists and songs
artist_songs <- tribble(
 ~artist, ~track,
 "J. Cole", "Motiv8",
 "Andrew Bird", "Anonanimal"
)
artist_songs \%>\%
 add_genius(artist, track, type = "lyrics")
}

}
