% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_playlist.R
\name{add_to_playlist}
\alias{add_to_playlist}
\title{Add tracks to a playlist}
\usage{
add_to_playlist(username, playlist_id, tracks, position = NULL,
  auth_code = get_spotify_authorization_code())
}
\arguments{
\item{username}{String of Spotify username. Can be found within the Spotify App}

\item{playlist_id}{The id of the playlist to add the tracks to. Can be obtained via either create_playlist() or get_playlists()}

\item{tracks}{A character vector of track uris}

\item{position}{(Optional) The position to insert the tracks, a zero-based index. If omitted, the tracks will be appended to the playlist}

\item{auth_code}{Authorization code with proper scopes. Calls get_spotify_authorization_code() by default.}
}
\description{
This function adds new tracks to a user's playlist.
}
\examples{
\dontrun{
hello_fromr <- create_playlist(username = 'rstats', playlist_name =  'hello from spotifyr!')
my_top_tracks <- get_my_top_tracks(time_range = 'long_term')
my_added_tracks <- add_to_playlist(username = 'rstats', playlist_id = hello_fromr$id, tracks =  my_top_tracks$track_uri)
}
}
\keyword{username}
