% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{auth}
\alias{auth}
\title{Building a function to auth the user}
\usage{
auth(clientID, secret)
}
\arguments{
\item{clientID}{char spotify clientID}

\item{secret}{char spotify secret}
}
\value{
a response
}
\description{
This function auth the user
}
\details{
This function authenticate the user
}
\author{
Yumeng Li, Mattias Karlsson, Ashraf Sarhan
}
\references{
\url{https://developer.spotify.com/web-api/authorization-guide/}
}
\seealso{
\code{\link{auth}}
}

