% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Authorize.R
\name{authorize_user}
\alias{authorize_user}
\alias{authorize_app}
\title{Get Tokens}
\usage{
authorize_user()

authorize_app()
}
\description{
Get Tokens
}
\section{Functions}{
\itemize{
\item \code{authorize_app}: 
}}

