% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Batch.R
\name{batch_col}
\alias{batch_col}
\title{Batch a tibble column}
\usage{
batch_col(x, col, n)
}
\arguments{
\item{x}{A tbl}

\item{col}{Column to batch}

\item{n}{Rows in each group}
}
\description{
Batch a tibble column
}
