% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{check_my_saved_albums}
\alias{check_my_saved_albums}
\title{Check User's Saved Albums}
\usage{
check_my_saved_albums(ids, authorization = get_spotify_authorization_code())
}
\arguments{
\item{ids}{Required. A comma-separated list of the \href{https://developer.spotify.com/documentation/web-api/#spotify-uris-and-ids}{Spotify IDs} for the albums. Maximum: 50 IDs.}

\item{authorization}{Required. A valid access token from the Spotify Accounts service. See the \href{https://developer.spotify.com/documentation/general/guides/authorization-guide/}{Web API authorization Guide} for more details. Defaults to \code{spotifyr::get_spotify_authorization_code()}. The access token must have been issued on behalf of the current user.}
}
\value{
Returns a data frame of results containing album id and saved status. See \url{https://developer.spotify.com/documentation/web-api/reference/library/check-users-saved-albums/} for more information.
}
\description{
Check if one or more albums is already saved in the current Spotify user’s ‘Your Music’ library.
}
