% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playlists.R
\name{create_playlist}
\alias{create_playlist}
\title{Create playlists}
\usage{
create_playlist(username, playlist_name,
  auth_code = get_spotify_authorization_code())
}
\arguments{
\item{username}{String of Spotify username. Can be found within the Spotify App}

\item{playlist_name}{Character vector of the new playlist name}

\item{auth_code}{Authorization code with proper scopes. Calls get_spotify_authorization_code() by default.}
}
\description{
This function creates a new playlist for a given Spotify username and returns a dataframe of metadata about the new playlist.
}
\examples{
\dontrun{
hello_fromr <- create_playlist(username = 'rstats', playlist_name =  'hello from spotifyr!')
}
}
\keyword{username}
