% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dedupe_album_names}
\alias{dedupe_album_names}
\title{Remove duplicate album names}
\usage{
dedupe_album_names(df, album_name_col = "album_name",
  album_release_year_col = "album_release_year")
}
\arguments{
\item{df}{Dataframe with album name}

\item{album_name_col}{String of field name containing album names}

\item{album_release_year_col}{String of field name containing album release year}
}
\description{
Use fuzzy matching to remove duplicate album names (including reissues, remasters, etc)
}
