% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/follow.R
\name{follow}
\alias{follow}
\title{Follow Artists or Users
Add the current user as a follower of one or more artists or other Spotify users.}
\usage{
follow(ids, type = c("artist", "user"), ...)
}
\description{
For more information: https://developer.spotify.com/web-api/follow-artists-users/
}

