% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/follow.R
\name{following}
\alias{following}
\title{Check if Current User Follows Artists or Users
Check to see if the current user is following one or more artists or other Spotify users}
\usage{
following(ids, type = c("artist", "user"), ...)
}
\description{
For more information: https://developer.spotify.com/web-api/check-current-user-follows/
}

