% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/follow.R
\name{following_playlist}
\alias{following_playlist}
\title{Check if Users Follow a Playlist
Check to see if one or more Spotify users are following a specified playlist.}
\usage{
following_playlist(owner_id, playlist_id, ids, ...)
}
\description{
For more information: https://developer.spotify.com/web-api/check-user-following-playlist/
}

