% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geniusR.R
\name{gen_album_url}
\alias{gen_album_url}
\title{Create Genius Album url}
\usage{
gen_album_url(artist = NULL, album = NULL)
}
\arguments{
\item{artist}{The quoted name of the artist. Spelling matters, capitalization does not.}

\item{album}{The quoted name of the album Spelling matters, capitalization does not.}
}
\description{
Creates a string containing the url to an album tracklist on Genius.com. The function is used internally to `genius_tracklist()`.
}
\examples{
\dontrun{
gen_album_url(artist = "Pinegrove", album = "Cardinal")
}
}
