% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geniusR.R
\name{gen_song_url}
\alias{gen_song_url}
\title{Create Genius url}
\usage{
gen_song_url(artist = NULL, song = NULL)
}
\arguments{
\item{artist}{The quoted name of the artist. Spelling matters, capitalization does not.}

\item{song}{The quoted name of the song. Spelling matters, capitalization does not.}
}
\description{
Generates the url for a song given an artist and a song title. This function is used internally within the `genius_lyrics()` function.
}
\examples{
gen_song_url(artist = "Kendrick Lamar", song = "HUMBLE")
gen_song_url("Margaret glaspy", "Memory Street")

}
