% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geniusR.R
\name{genius_lyrics}
\alias{genius_lyrics}
\title{Retrieve song lyrics from Genius.com}
\usage{
genius_lyrics(artist = NULL, song = NULL, info = "title")
}
\arguments{
\item{artist}{The quoted name of the artist. Spelling matters, capitalization does not.}

\item{song}{The quoted name of the song. Spelling matters, capitalization does not.}

\item{info}{Default `"title"`, returns the track title. Set to `"simple"` for only lyrics, `"artist"` for the lyrics and artist, or `"all"` to return the lyrics, artist, and title.}
}
\description{
Retrieve the lyrics of a song with supplied artist and song name.
}
\examples{
genius_lyrics(artist = "Margaret Glaspy", song = "Memory Street")
genius_lyrics(artist = "Kendrick Lamar", song = "Money Trees")
genius_lyrics("JMSN", "Drinkin'")

}
