% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geniusR.R
\name{genius_tracklist}
\alias{genius_tracklist}
\title{Create a tracklist of an album}
\usage{
genius_tracklist(artist = NULL, album = NULL)
}
\arguments{
\item{artist}{The quoted name of the artist. Spelling matters, capitalization does not.}

\item{album}{The quoted name of the album Spelling matters, capitalization does not.}
}
\description{
Creates a `tibble` containing all track titles for a given artist and album. This function is used internally in `genius_album()`.
}
\examples{
\dontrun{
genius_tracklist(artist = "Andrew Bird", album = "Noble Beast")
}
}
