% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/albums.R
\name{get_album}
\alias{get_album}
\title{Get an Album
Get Spotify catalog information for a single album}
\usage{
get_album(id, ...)
}
\arguments{
\item{id}{Required. The Spotify ID for the album

For more information: https://developer.spotify.com/web-api/get-album/}
}
\description{
Get an Album
Get Spotify catalog information for a single album
}

