% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_album_data.R
\name{get_album_data}
\alias{get_album_data}
\title{Retrieve artist discography with song lyrics and audio info}
\usage{
get_album_data(artist, albums = character(), parallelize = FALSE,
  future_plan = "multiprocess", access_token = get_spotify_access_token())
}
\arguments{
\item{artist}{The quoted name of the artist. Spelling matters, capitalization does not.}

\item{albums}{A character vector of album names. Spelling matters, capitalization does not}

\item{parallelize}{Boolean determining to run in parallel or not. Defaults to \code{FALSE}.}

\item{future_plan}{String determining how `future()`s are resolved when `parallelize == TRUE`. Defaults to \code{multiprocess}.}
}
\description{
Retrieve the entire discography of an artist with the lyrics of each song and the associated audio information. Returns the song data as a nested tibble. This way we can easily see each album, artist, and song title before expanding our data.
}
\examples{
get_album_data("Wild child", "Expectations")

}
