% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/albums.R
\name{get_album_tracks}
\alias{get_album_tracks}
\title{Get tracks from one or more albums on Spotify}
\usage{
get_album_tracks(albums, access_token = get_spotify_access_token(),
  parallelize = FALSE, future_plan = "multiprocess")
}
\arguments{
\item{albums}{A character vector containing the ids of one or several albums to retrieve information for OR a dataframe containing a column `album_uri`, corresponding to Spotify Album URIs (Can be output from spotifyr::get_artist_albums())}

\item{access_token}{Spotify Web API token. Defaults to spotifyr::get_spotify_access_token()}

\item{parallelize}{Boolean determining to run in parallel or not. Defaults to \code{FALSE}.}

\item{future_plan}{String determining how `future()`s are resolved when `parallelize == TRUE`. Defaults to \code{multiprocess}.}
}
\description{
This function returns tracks from a dataframe of albums on Spotify
}
\examples{
\dontrun{
albums <- get_artist_albums('radiohead')
get_album_tracks(albums)
}
}
\keyword{album}
\keyword{tracks}
