% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_albums.R
\name{get_albums}
\alias{get_albums}
\title{Get album metadata}
\usage{
get_albums(albums = character(), access_token = get_spotify_access_token())
}
\arguments{
\item{albums}{A character vector containing the ids of one or several albums to retrieve information for}

\item{access_token}{Spotify Web API token. Defaults to spotifyr::get_spotify_access_token()}
}
\value{
Returns a data frame of results containing album data. See \url{https://developer.spotify.com/documentation/web-api/reference/albums/get-album/} for more information.
}
\description{
Get album metadata
}
\examples{

met_albums <- get_artist_albums(artist = "Metallica")
met_meta <- get_album_meta(albums = met_albums$album_uri)
}
