% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_artist_albums.R
\name{get_artist_albums}
\alias{get_artist_albums}
\title{Get Artist Albums}
\usage{
get_artist_albums(artist_uri, studio_albums_only = TRUE,
  access_token = get_spotify_access_token())
}
\arguments{
\item{artist_uri}{String identifier for an artist on Spotify. Can be found within the Spotify app or with spotifyr::get_artists()}

\item{studio_albums_only}{Logical for whether to remove album types "single" and "compilation" and albums with mulitple artists. Defaults to \code{TRUE}}

\item{access_token}{Spotify Web API token. Defaults to spotifyr::get_spotify_access_token()}
}
\description{
This function returns an artist's discography on Spotify
}
\examples{
\dontrun{
artists <- get_artists('radiohead')
albums <- get_artist_albums(artists$artist_uri[1])
}
}
\keyword{albums}
