% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artists.R
\name{get_artist_toptracks}
\alias{get_artist_toptracks}
\title{Get an Artist’s Top Tracks
Get Spotify catalog information about an artist’s top tracks by country.
filter and sort the response.}
\usage{
get_artist_toptracks(id, country = "US")
}
\arguments{
\item{id}{Required. The Spotify ID for the artist}
}
\description{
Get an Artist’s Top Tracks
Get Spotify catalog information about an artist’s top tracks by country.
filter and sort the response.
}
\examples{
set_tokens()
get_artist_toptracks('1vCWHaC5f2uS3yhpwWbIA6')
get_artist_toptracks('1vCWHaC5f2uS3yhpwWbIA6', 'DE')
}
\references{
\href{https://developer.spotify.com/web-api/get-artists-top-tracks/}{API documentation}
}
