% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artists.R
\name{get_artists}
\alias{get_artists}
\title{Get Spotify catalog information for multiple artists identified by their Spotify IDs.}
\usage{
get_artists(ids, Authorization = get_spotify_access_token())
}
\arguments{
\item{ids}{Required. A character vector of the \href{https://developer.spotify.com/documentation/web-api/#spotify-uris-and-ids}{Spotify IDs} for the artists. Maximum: 50 IDs.}

\item{Authorization}{Required. A valid access token from the Spotify Accounts service. See the \href{Web API Authorization Guide}{https://developer.spotify.com/documentation/general/guides/authorization-guide/} for more details. Defaults to \code{spotifyr::get_spotify_access_token()}}
}
\value{
Returns a data frame of results containing artist data. See \url{https://developer.spotify.com/documentation/web-api/reference/artists/get-several-artists/} for more information.
}
\description{
Get Spotify catalog information for multiple artists identified by their Spotify IDs.
}
\examples{

}
