% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/profiles.R
\name{get_current_user_profile}
\alias{get_current_user_profile}
\title{Get Current User’s Profile
Get detailed profile information about the current user (including the current user’s username).
For more information: https://developer.spotify.com/web-api/get-playlist/}
\usage{
get_current_user_profile()
}
\description{
Get Current User’s Profile
Get detailed profile information about the current user (including the current user’s username).
For more information: https://developer.spotify.com/web-api/get-playlist/
}

