% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.R
\name{get_currently_playing_track}
\alias{get_currently_playing_track}
\title{Get the User’s Currently Playing Track
Get the object currently being played on the user’s Spotify account.}
\usage{
get_currently_playing_track()
}
\description{
Get the User’s Currently Playing Track
Get the object currently being played on the user’s Spotify account.
}
\examples{
set_tokens()
user_auth()
get_currently_playing_track()
}
\references{
\href{https://developer.spotify.com/web-api/get-the-users-currently-playing-track/}{API documentation}
}
