% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{get_featured_playlists}
\alias{get_featured_playlists}
\title{Get a List of Featured Playlists
Get a list of Spotify featured playlists (shown, for example, on a Spotify player’s “Browse” tab).}
\usage{
get_featured_playlists(...)
}
\description{
Get a List of Featured Playlists
Get a list of Spotify featured playlists (shown, for example, on a Spotify player’s “Browse” tab).
}
\examples{
set_tokens()
get_featured_playlists()
}
\references{
\href{https://developer.spotify.com/web-api/get-list-featured-playlists/}{API documentation}
}
