% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tracks.R
\name{get_features}
\alias{get_features}
\alias{get_details}
\title{Get Track Data}
\usage{
get_features(ids)

get_details(ids)
}
\arguments{
\item{ids}{Vector of track IDs}
}
\description{
Get Track Data
}
\section{Functions}{
\itemize{
\item \code{get_details}: 
}}

