% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genre_artists.R
\name{get_genre_artists}
\alias{get_genre_artists}
\title{Get artists by genre}
\usage{
get_genre_artists(genre = character(), offset = 0, limit = 20,
  access_token = get_spotify_access_token())
}
\arguments{
\item{genre}{String of genre name to search for}

\item{offset}{Integer indicating the offset of the first artist to return. Defaults to 0 (Spotify's API default value).}

\item{limit}{Integer indicating the max number of artists to return. Defaults to 20 (Spotify's API default value), max of 50.}

\item{access_token}{Spotify Web API token. Defaults to spotifyr::get_spotify_access_token()}
}
\value{
Returns a data frame of results containing artist data for the given genre.
}
\description{
Get artists by genre
}
\examples{
wonky_artists <- get_genre_artists("wonky")
}
