% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_label_albums.R
\name{get_label_albums}
\alias{get_label_albums}
\title{Get albums by label}
\usage{
get_label_albums(label = character(), album_types = "album", offset = 0,
  limit = 20, access_token = get_spotify_access_token())
}
\arguments{
\item{label}{String of label name to search for}

\item{album_types}{Character vector of album types to include. Valid values are "album", "single", "appears_on", and "compilation". Defaults to "album".}

\item{offset}{Integer indicating the offset of the first album to return. Defaults to 0 (Spotify's API default value).}

\item{limit}{Integer indicating the max number of album to return. Defaults to 20 (Spotify's API default value), max of 50.}

\item{access_token}{Spotify Web API token. Defaults to spotifyr::get_spotify_access_token()}
}
\value{
Returns a data frame of results containing album data for the given label.
}
\description{
Get albums by label
}
\examples{
brainfeeder_albums <- get_label_albums("brainfeeder")
}
