% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{get_label_artists}
\alias{get_label_artists}
\title{Get artists by label}
\usage{
get_label_artists(label = character(), offset = 0, limit = 20,
  access_token = get_spotify_access_token())
}
\arguments{
\item{label}{String of label name to search for}

\item{offset}{Integer indicating the offset of the first artist to return. Defaults to 0 (Spotify's API default value).}

\item{limit}{Integer indicating the max number of artists to return. Defaults to 20 (Spotify's API default value), max of 50.}

\item{access_token}{Spotify Web API token. Defaults to spotifyr::get_spotify_access_token()}
}
\value{
Returns a data frame of results containing artist data for the given label.
}
\description{
Get artists by label
}
\examples{
brainfeeder_artists <- get_label_artists("brainfeeder")
}
