% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{get_my_currently_playing}
\alias{get_my_currently_playing}
\title{Get currently playing track for current user}
\usage{
get_my_currently_playing(auth_code = get_spotify_authorization_code())
}
\arguments{
\item{auth_code}{Authorization code with proper scopes. Calls get_spotify_authorization_code() by default.}

\item{simplify}{Boolean deciding whether to simplify JSON object into a dataframe with pre-selected fields for analysis. Defaults to \code{FALSE}.}
}
\description{
Get currently playing track for current user
}
\examples{
\dontrun{
get_my_currently_playing()
}
}
\keyword{auth}
