% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{get_my_recently_played}
\alias{get_my_recently_played}
\title{Get recently played tracks for current user}
\usage{
get_my_recently_played(limit = 50,
  auth_code = get_spotify_authorization_code())
}
\arguments{
\item{limit}{Integer indicating the number of tracks to return. Default: 50. Minimum: 1. Maximum: 50}

\item{auth_code}{Authorization code with proper scopes. Calls get_spotify_authorization_code() by default.}

\item{after}{A Unix timestamp in milliseconds. Returns all items after (but not including) this cursor position. If after is specified, before must not be specified.}

\item{before}{A Unix timestamp in milliseconds. Returns all items before (but not including) this cursor position. If before is specified, after must not be specified.}
}
\description{
Get recently played tracks for current user
}
\examples{
\dontrun{
get_my_recently_played()
}
}
\keyword{auth}
