% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{get_my_saved_shows}
\alias{get_my_saved_shows}
\title{Get User's Saved Shows}
\usage{
get_my_saved_shows(
  limit = 20,
  offset = 0,
  authorization = get_spotify_authorization_code(),
  include_meta_info = FALSE
)
}
\arguments{
\item{limit}{Optional. \cr
Maximum number of shows to return. \cr
Default: 20 \cr
Minimum: 1 \cr
Maximum: 50 \cr}

\item{offset}{Optional. \cr
The index of the first show to return. \cr
Default: 0 (the first object). Use with \code{limit} to get the next set of shows.}

\item{authorization}{Required. \cr
A valid access token from the Spotify Accounts service. See the \href{https://developer.spotify.com/documentation/general/guides/authorization-guide/}{Web API authorization Guide} for more details. Defaults to \code{spotifyr::get_spotify_authorization_code()}. The access token must have been issued on behalf of the current user.}

\item{include_meta_info}{Optional. \cr
Boolean indicating whether to include full result, with meta information such as \code{"total"}, and \code{"limit"}. Defaults to \code{FALSE}.}
}
\value{
Returns a data frame of results containing saved shows information. See \url{https://developer.spotify.com/documentation/web-api/reference/library/get-users-saved-shows/} for more information.
}
\description{
Get a list of shows saved in the current Spotify user’s library. Optional parameters can be used to limit the number of shows returned.
}
