% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{get_my_top_artists}
\alias{get_my_top_artists}
\title{Get top artists for Current User}
\usage{
get_my_top_artists(time_range = "medium_term", limit = 50,
  offset = 0, auth_code = get_spotify_authorization_code())
}
\arguments{
\item{time_range}{String indicating the time frame over which the affinities are computed. Valid values: long_term (calculated from several years of data and including all new data as it becomes available), medium_term (approximately last 6 months), short_term (approximately last 4 weeks). Default: medium_term.}

\item{limit}{Integer indicating the number of entities to return. Default: 50. Minimum: 1. Maximum: 50}

\item{offset}{Integer indicating the index of the first entity to return. Default: 0 (i.e., the first artist). Use with limit to get the next set of entities.}

\item{auth_code}{Authorization code with proper scopes. Calls get_spotify_authorization_code() by default.}
}
\description{
Get top artists for current user
}
\examples{
\dontrun{
get_my_top_artists()
}
}
\keyword{auth}
