% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{get_new_releases}
\alias{get_new_releases}
\title{Get a List of New Releases
Get a list of new album releases featured in Spotify (shown, for example, on a Spotify player’s “Browse” tab).}
\usage{
get_new_releases(country, ...)
}
\arguments{
\item{country}{Required if you want any results.
A country: an ISO 3166-1 alpha-2 country code.
Provide this parameter if you want the list of
returned items to be relevant to a particular country.}
}
\description{
Get a List of New Releases
Get a list of new album releases featured in Spotify (shown, for example, on a Spotify player’s “Browse” tab).
}
\examples{
get_new_releases(country = 'US')
}
\references{
\href{https://developer.spotify.com/web-api/get-list-new-releases/}{API documentation}
}
