% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_playlist_audio_features.R
\name{get_playlist_audio_features}
\alias{get_playlist_audio_features}
\title{Get features and popularity for all of a given set of playlists on Spotify}
\usage{
get_playlist_audio_features(username, playlist_uris,
  access_token = get_spotify_access_token())
}
\arguments{
\item{username}{String of Spotify username. Can be found on the Spotify app. (See http://rcharlie.net/sentify/user_uri.gif for example)}

\item{playlist_uris}{Character vector of Spotify playlist uris associated with the given \code{username}. Can be found within the Spotify App}

\item{access_token}{Spotify Web API token. Defaults to spotifyr::get_spotify_access_token()}
}
\description{
This function returns the popularity and audio features for every song for a given set of playlists on Spotify
}
\examples{
\dontrun{
playlist_username <- 'spotify'
playlist_uris <- c('37i9dQZF1E9T1oFsQFg98K', '37i9dQZF1CyQNOI21QVf3p')
my_playlist_audio_features <- get_playlist_audio_features('spotify', playlist_uris)
}
}
\keyword{audio}
\keyword{features}
\keyword{playlists}
\keyword{track}
