% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_playlists.R
\name{get_playlists}
\alias{get_playlists}
\title{Get playlists}
\usage{
get_playlists(username, playlist_uris,
  access_token = get_spotify_access_token())
}
\arguments{
\item{username}{String of Spotify username. Can be found within the Spotify App}

\item{playlist_uris}{Character vector of Spotify playlist uris associated with the given \code{username}. Can be found within the Spotify App}

\item{access_token}{Spotify Web API token. Defaults to \code{spotifyr::get_spotify_access_token()}.}

\item{show_progress}{Boolean determining to show progress bar or not. Defaults to \code{FALSE}.}
}
\description{
This function returns a dataframe of playlists for a given Spotify username and a character vector of playlist uris
}
\examples{
\dontrun{
playlist_username <- 'spotify'
playlist_uris <- c('37i9dQZF1E9T1oFsQFg98K', '37i9dQZF1CyQNOI21QVf3p')
my_playlist_audio_features <- get_playlist_audio_features('spotify', playlist_uris)
}
}
\keyword{username}
