% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_related_artists.R
\name{get_related_artists}
\alias{get_related_artists}
\title{Get Related Artists}
\usage{
get_related_artists(artist = NULL, return_closest_artist = TRUE,
  access_token = get_spotify_access_token())
}
\arguments{
\item{artist}{String of artist name or Spotify artist URI}

\item{return_closest_artist}{Boolean for selecting the artist result with the closest match on Spotify's Search endpoint. Defaults to \code{TRUE}.}

\item{access_token}{Spotify Web API token. Defaults to \code{spotifyr::get_spotify_access_token()}.}
}
\description{
This function searches Spotify's library for artists by name or Spotify URI and returns related artists using Spotify's "Related Artists" API endpoint.
}
\examples{
\dontrun{
get_related_artists('radiohead')

purrr::map_df(vector_of_artist_uris, function(this_artist_uri) {
    get_related_artists(artist = this_artist_uri)
})
}
}
\keyword{artists}
\keyword{related}
