% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{get_saved_tracks}
\alias{get_saved_tracks}
\title{Get a User’s Saved Tracks
Get a list of the songs saved in the current Spotify user’s “Your Music” library.}
\usage{
get_saved_tracks(...)
}
\description{
Get a User’s Saved Tracks
Get a list of the songs saved in the current Spotify user’s “Your Music” library.
}
\examples{
set_tokens()
user_auth()
get_saved_tracks()
}
\references{
\href{https://developer.spotify.com/web-api/get-users-saved-tracks/}{API documentation}
}
