% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{get_spotify_access_token}
\alias{get_spotify_access_token}
\title{Get Spotify Access Token}
\usage{
get_spotify_access_token(
  client_id = Sys.getenv("SPOTIFY_CLIENT_ID"),
  client_secret = Sys.getenv("SPOTIFY_CLIENT_SECRET")
)
}
\arguments{
\item{client_id}{Defaults to System Environment variable "SPOTIFY_CLIENT_ID"}

\item{client_secret}{Defaults to System Environment variable "SPOTIFY_CLIENT_SECRET"}
}
\value{
The Spotify Access Token of the current user or a meaningful error message.
}
\description{
This function creates a Spotify access token.
}
\examples{
\donttest{
token <- get_spotify_access_token()
}
}
\seealso{
Other authentication functions: 
\code{\link{get_spotify_authorization_code}()}
}
\concept{authentication functions}
