% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Top-Tracks.R
\name{get_top_tracks}
\alias{get_top_tracks}
\title{Get Your Top Tracks}
\usage{
get_top_tracks(range = "medium_term")
}
\arguments{
\item{range}{Timespan to cover: long_term (years), medium_term (6 months), short_term (4 weeks)}
}
\description{
Get Your Top Tracks
}
