% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_track.R
\name{get_track}
\alias{get_track}
\title{Get track uri:s from a string search on Spotify}
\usage{
get_track(track_name, artist_name = NULL, album_name = NULL,
  access_token = get_spotify_access_token())
}
\arguments{
\item{track_name}{A string with track name}

\item{artist_name}{Optional. A string with artist name}

\item{album_name}{Optional. A string with album name}

\item{access_token}{Spotify Web API token. Defaults to spotifyr::get_spotify_access_token()}
}
\description{
This function takes a string and returns a data frame with track information
from Spotify's search endpoint
}
\examples{
\dontrun{
##### Get track uri for Radiohead - Kid A
kid_a <- get_track(artist_name = "Radiohead", track_name = "Kid A")
}
}
\keyword{search}
\keyword{string}
\keyword{track}
\keyword{uri}
