% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_track_audio_analysis.R
\name{get_track_audio_analysis}
\alias{get_track_audio_analysis}
\title{Get track audio analysis by URI}
\usage{
get_track_audio_analysis(track_uri, access_token = get_spotify_access_token())
}
\arguments{
\item{track_uri}{A string with a track URI}

\item{access_token}{Spotify Web API token. Defaults to spotifyr::get_spotify_access_token()}
}
\description{
This function takes a track URI and returns a list containing an audio analysis object
from Spotify's audio analysis endpoint
}
\examples{
\dontrun{
##### Get track uri for Radiohead - Kid A
kid_a <- get_tracks(artist_name = "Radiohead", track_name = "Kid A", return_closest_track = TRUE)
kid_a_audio_analysis <- get_track_audio_analysis(kid_a$track_uri)
}
}
\keyword{search}
\keyword{string}
\keyword{track}
\keyword{uri}
