% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks.R
\name{get_track_audio_features}
\alias{get_track_audio_features}
\title{Get audio feature information for a single track identified by its unique Spotify ID.}
\usage{
get_track_audio_features(id, Authorization = get_spotify_access_token())
}
\arguments{
\item{id}{The \href{https://developer.spotify.com/documentation/web-api/#spotify-uris-and-ids}{Spotify ID} for the track.}

\item{Authorization}{Required. A valid access token from the Spotify Accounts service. See the \href{Web API Authorization Guide}{https://developer.spotify.com/documentation/general/guides/authorization-guide/} for more details. Defaults to \code{spotifyr::get_spotify_access_token()}}
}
\value{
Returns a data frame of results containing track audio features data. See \url{https://developer.spotify.com/documentation/web-api/reference/tracks/get-audio-analysis//} for more information.
}
\description{
Get audio feature information for a single track identified by its unique Spotify ID.
}
\examples{

}
