% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks.R
\name{get_tracks}
\alias{get_tracks}
\title{Get track uris from a string search on Spotify}
\usage{
get_tracks(track_name, artist_name = NULL, album_name = NULL,
  return_closest_track = FALSE, access_token = get_spotify_access_token())
}
\arguments{
\item{track_name}{A string with track name}

\item{artist_name}{Optional. A string with artist name}

\item{album_name}{Optional. A string with album name}

\item{return_closest_track}{Optional. A string with album name}

\item{access_token}{Spotify Web API token. Defaults to spotifyr::get_spotify_access_token()}
}
\description{
This function takes a string and returns a data frame with track information
from Spotify's search endpoint
}
\examples{
\dontrun{
##### Get track uri for Radiohead - Kid A
kid_a <- get_tracks(artist_name = "Radiohead", track_name = "Kid A", return_closest_track = TRUE)
}
}
\keyword{search}
\keyword{string}
\keyword{track}
\keyword{uri}
