% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auth.R
\name{get_user_code}
\alias{get_user_code}
\title{Get user code for Authorization Code user code
Lauches Selenium Webbrowser to handle process}
\usage{
get_user_code()
}
\description{
Get user code for Authorization Code user code
Lauches Selenium Webbrowser to handle process
}

