% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.R
\name{get_user_devices}
\alias{get_user_devices}
\title{Get a User’s Available Devices
Get information about a user’s available devices.}
\usage{
get_user_devices()
}
\description{
Get a User’s Available Devices
Get information about a user’s available devices.
}
\examples{
set_tokens()
user_auth()
get_user_devices()
}
\references{
\href{https://developer.spotify.com/web-api/get-a-users-available-devices/}{API documentation}
}
