% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/follow.R
\name{get_user_followed_artists}
\alias{get_user_followed_artists}
\title{Get User’s Followed Artists
Get the current user’s followed artists.}
\usage{
get_user_followed_artists(...)
}
\description{
For more information: https://developer.spotify.com/web-api/get-followed-artists/
}

