% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.R
\name{get_user_playback}
\alias{get_user_playback}
\title{Get Information About The User’s Current Playback
Get information about the user’s current playback state, including track, track progress, and active device.}
\usage{
get_user_playback()
}
\description{
Get Information About The User’s Current Playback
Get information about the user’s current playback state, including track, track progress, and active device.
}
\examples{
set_tokens()
user_auth()
get_user_playback()
}
\references{
\href{https://developer.spotify.com/web-api/get-information-about-the-users-current-playback/}{API documentation}
}
