% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_playlists.R
\name{get_user_playlists}
\alias{get_user_playlists}
\title{Get user playlists}
\usage{
get_user_playlists(username, access_token = get_spotify_access_token())
}
\arguments{
\item{username}{String of Spotify username. Can be found within the Spotify App}

\item{access_token}{Spotify Web API token. Defaults to spotifyr::get_spotify_access_token()}
}
\description{
This function returns a dataframe of playlists for a given Spotify username
}
\examples{
get_user_playlists('barackobama')
}
\keyword{username}
