% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalization.R
\name{get_user_recently_played}
\alias{get_user_recently_played}
\title{Get Current User’s Recently Played Tracks
Get tracks from the current user’s recently played tracks.}
\usage{
get_user_recently_played()
}
\description{
Get Current User’s Recently Played Tracks
Get tracks from the current user’s recently played tracks.
}
\examples{
set_tokens()
user_auth()
get_user_recently_played()
}
\references{
\href{https://developer.spotify.com/web-api/web-api-personalization-endpoints/get-recently-played/}{API documentation}
}
