% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auth.R
\name{get_user_token}
\alias{get_user_token}
\title{Using the user_code, generates tokens for user code}
\usage{
get_user_token(user_code)
}
\description{
Using the user_code, generates tokens for user code
}

