% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalization.R
\name{get_user_top}
\alias{get_user_top}
\title{Get a User’s Top Artists and Tracks
Get the current user’s top artists or tracks based on calculated affinity.}
\usage{
get_user_top(type = c("artists", "tracks"))
}
\arguments{
\item{type}{}
}
\description{
Get a User’s Top Artists and Tracks
Get the current user’s top artists or tracks based on calculated affinity.
}
\examples{
set_tokens()
user_auth()
get_user_top(type='artists')
get_user_top(type='tracks')
}
\references{
\href{https://developer.spotify.com/web-api/get-users-top-artists-and-tracks/}{API documentation}
}
