% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks.R
\name{getTracks}
\alias{getTracks}
\title{Building a function to access to the specific tracks}
\usage{
getTracks(trackIDs, market, access_token)
}
\arguments{
\item{market}{a charater of the market country}

\item{access_token}{a charater of the access token}

\item{artistID}{a character of the ID of spotify tracks}
}
\value{
a response
}
\description{
This function is building the access to the specific tracks
}
\details{
This function is to get the acesss to the specific tracks
}
\author{
Yumeng Li, Mattias Karlsson, Ashraf Sarhan
}
\references{
\url{https://developer.spotify.com/web-api/authorization-guide/}
}
\seealso{
\code{\link{auth}}
}

