% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Batchr.R
\name{group_n}
\alias{group_n}
\title{Batch a tibble column}
\usage{
group_n(x, col, n)
}
\arguments{
\item{x}{A tbl}

\item{col}{Column to batch}

\item{n}{Rows in each group}
}
\description{
Batch a tibble column
}
