% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_playlists.R
\name{parse_playlist_list_to_df}
\alias{parse_playlist_list_to_df}
\title{Get user playlists}
\usage{
parse_playlist_list_to_df(playlist_list)
}
\arguments{
\item{artist_name}{String of Spotify username. Can be found within the Spotify App}

\item{access_token}{Spotify Web API token. Defaults to spotifyr::get_spotify_access_token()}
}
\description{
This function returns a dataframe of playlists for a given Spotify username
}
\examples{
get_user_playlists('barackobama')
}
\keyword{username}
