% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.R
\name{pause_user_playback}
\alias{pause_user_playback}
\title{Pause a User’s Playback
Pause playback on the user’s account.}
\usage{
pause_user_playback(device_id = NULL)
}
\arguments{
\item{device_id}{Optional. The id of the device this command is targeting. If not supplied, the user's currently active device is the target}
}
\description{
Pause a User’s Playback
Pause playback on the user’s account.
}
\examples{
set_tokens()
user_auth()
pause_user_playback()
}
\references{
\href{https://developer.spotify.com/web-api/pause-a-users-playback/}{API documentation}
}
