% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geniusR.R
\name{prep_info}
\alias{prep_info}
\title{Prepares input strings for `gen_song_url()`}
\usage{
prep_info(input)
}
\arguments{
\item{input}{Either artist, song, or album, function input.}
}
\description{
Applies a number of regular expressions to prepare the input to match Genius url format
}
