% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playlists.R
\name{print.playlist}
\alias{print.playlist}
\title{Print method for playlist object}
\usage{
\method{print}{playlist}(x, ...)
}
\arguments{
\item{x}{A playlist object generated from \code{get_playlist()}.}

\item{...}{Unused.}
}
\description{
Print method for playlist object
}
