% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{refresh_user_token}
\alias{refresh_user_token}
\title{Refresh your tokens}
\usage{
refresh_user_token(token = NULL)
}
\arguments{
\item{token}{default = NULL.  If no token is given, it will look for refresh_token in global environment}
}
\description{
Refresh your tokens
}
