% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/albums.R
\name{remove_albums_from_my_library}
\alias{remove_albums_from_my_library}
\title{Remove one or more albums from current user's collection}
\usage{
remove_albums_from_my_library(
  ids,
  authorization = get_spotify_authorization_code()
)
}
\arguments{
\item{ids}{Required. A character vector of the \href{https://developer.spotify.com/documentation/web-api/#spotify-uris-and-ids}{Spotify IDs} for the albums. Maximum: 50 IDs.}

\item{authorization}{Required. A valid access token from the Spotify Accounts service. See the \href{https://developer.spotify.com/documentation/general/guides/authorization-guide/}{Web API authorization Guide} for more details. Defaults to \code{spotifyr::get_spotify_access_token()}}
}
\description{
Remove one or more albums from current user's collection
}
