% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{remove_my_saved_tracks}
\alias{remove_my_saved_tracks}
\title{Remove User's Saved Tracks}
\usage{
remove_my_saved_tracks(
  ids,
  authorization = get_spotify_authorization_code(),
  echo = FALSE
)
}
\arguments{
\item{ids}{Required. \cr
A comma-separated list of the \href{https://developer.spotify.com/documentation/web-api/#spotify-uris-and-ids}{Spotify IDs} for the albums. Maximum: 50 IDs.}

\item{authorization}{Required. \cr
A valid access token from the Spotify Accounts service. See the \href{https://developer.spotify.com/documentation/general/guides/authorization-guide/}{Web API authorization Guide} for more details. Defaults to \code{spotifyr::get_spotify_authorization_code()}. The access token must have been issued on behalf of the current user.}

\item{echo}{Optional.\cr
Boolean indicating whether to return the response or work silently.}
}
\description{
Remove one or more tracks from the current user’s ‘Your Music’ library. \cr
Changes to a user’s saved tracks may not be visible in other Spotify applications immediately.
}
