% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{save_shows_to_library}
\alias{save_shows_to_library}
\title{Save Shows for Current User}
\usage{
save_shows_to_library(
  ids,
  authorization = get_spotify_authorization_code(),
  echo = FALSE
)
}
\arguments{
\item{ids}{Required. \cr
A comma-separated list of the \href{https://developer.spotify.com/documentation/web-api/#spotify-uris-and-ids}{Spotify IDs} for the albums. Maximum: 50 IDs.}

\item{authorization}{Required. \cr
A valid access token from the Spotify Accounts service. See the \href{https://developer.spotify.com/documentation/general/guides/authorization-guide/}{Web API authorization Guide} for more details. Defaults to \code{spotifyr::get_spotify_authorization_code()}. The access token must have been issued on behalf of the current user.}

\item{echo}{Optional.\cr
Boolean indicating whether to return the response or work silently.}
}
\value{
Returns a respinse status. See \url{https://developer.spotify.com/documentation/web-api/#response-status-codes} for more information.
}
\description{
Save one or more shows to current Spotify user’s library.
}
