% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{save_track}
\alias{save_track}
\title{Save Tracks for User
Save one or more tracks to the current user’s “Your Music” library}
\usage{
save_track(ids, ...)
}
\arguments{
\item{ids}{}
}
\description{
Save Tracks for User
Save one or more tracks to the current user’s “Your Music” library
}
\examples{
set_tokens()
user_auth()
save_track(ids=c('4iV5W9uYEdYUVa79Axb7Rh', '4iV5W9uYEdYUVa79Axb7Rh'))
save_track(ids='4iV5W9uYEdYUVa79Axb7Rh')
}
\references{
\href{https://developer.spotify.com/web-api/save-tracks-user/}{API documentation}
}
