% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks.R
\name{save_tracks_to_my_library}
\alias{save_tracks_to_my_library}
\title{Save one or more tracks to a user's collection}
\usage{
save_tracks_to_my_library(
  ids,
  authorization = get_spotify_authorization_code()
)
}
\arguments{
\item{ids}{The \href{https://developer.spotify.com/documentation/web-api/#spotify-uris-and-ids}{Spotify ID} for the tracks. Maximum: 50 IDs.}

\item{authorization}{Required. A valid access token from the Spotify Accounts service. See the \href{https://developer.spotify.com/documentation/general/guides/authorization-guide/}{Web API authorization guide} for more details. Defaults to \code{spotifyr::get_spotify_access_token()}}
}
\description{
Save one or more tracks to a user's collection
}
