% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/search.R
\name{search}
\alias{search}
\title{Get Spotify catalog information about artists, albums,
tracks or playlists that match a keyword string.}
\usage{
search(q, type, ...)
}
\arguments{
\item{q}{Required. The search query's keywords
(and optional field filters and operators),
for example q=roadhouse%20blues.}

\item{Required.}{A comma-separated list of item types
to search across. Valid types are: album, artist,
playlist, and track}
}
\description{
Get Spotify catalog information about artists, albums,
tracks or playlists that match a keyword string.
}

