% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.R
\name{seek_user_playback}
\alias{seek_user_playback}
\title{Seek To Position In Currently Playing Track
Seeks to the given position in the user’s currently playing track.}
\usage{
seek_user_playback(position_ms, device_id = NULL)
}
\arguments{
\item{position_ms}{Required. The position in milliseconds to seek to. Must be a positive number. Passing in a position that is greater than the length of the track will cause the player to start playing the next song.}

\item{device_id}{Optional. The id of the device this command is targeting. If not supplied, the user's currently active device is the target}
}
\description{
Seek To Position In Currently Playing Track
Seeks to the given position in the user’s currently playing track.
}
\examples{
set_tokens()
user_auth()
seek_user_playback(position_ms = 1000)
}
\references{
\href{https://developer.spotify.com/web-api/seek-to-position-in-currently-playing-track/}{API documentation}
}
