% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auth.R
\name{set_credentials}
\alias{set_credentials}
\title{Set credentials to be accessed by all functions}
\usage{
set_credentials(client_id, client_secret)
}
\description{
Set credentials to be accessed by all functions
}

