% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.R
\name{skip_my_playback}
\alias{skip_my_playback}
\title{Skips to next track in the user’s queue.}
\usage{
skip_my_playback(device_id = NULL,
  authorization = get_spotify_authorization_code())
}
\arguments{
\item{device_id}{Optional. The id of the device this command is targeting. If not supplied, the user’s currently active device is the target.}

\item{authorization}{Required. A valid access token from the Spotify Accounts service. See the \href{https://developer.spotify.com/documentation/general/guides/authorization-guide/}{Web API authorization Guide} for more details. Defaults to \code{spotifyr::get_spotify_access_token()}. The access token must have been issued on behalf of the current user. \cr
The access token must have the \code{user-modify-playback-state} scope authorized in order to control playback.}
}
\description{
Skips to next track in the user’s queue.
}
