% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.R
\name{skip_user_playback_next}
\alias{skip_user_playback_next}
\title{Skip User’s Playback To Next Track
Skips to next track in the user’s queue.}
\usage{
skip_user_playback_next(device_id = NULL)
}
\arguments{
\item{device_id}{Optional. The id of the device this command is targeting. If not supplied, the user's currently active device is the target}
}
\description{
Skip User’s Playback To Next Track
Skips to next track in the user’s queue.
}
\examples{
set_tokens()
user_auth()
skip_user_playback_next()
}
\references{
\href{https://developer.spotify.com/web-api/skip-users-playback-to-next-track/}{API documentation}
}
