% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.R
\name{skip_user_playback_previous}
\alias{skip_user_playback_previous}
\title{Skip User’s Playback To Previous Track
Skips to previous track in the user’s queue.}
\usage{
skip_user_playback_previous(device_id = NULL)
}
\arguments{
\item{device_id}{Optional. The id of the device this command is targeting. If not supplied, the user's currently active device is the target}
}
\description{
Skip User’s Playback To Previous Track
Skips to previous track in the user’s queue.
}
\examples{
set_tokens()
user_auth()
skip_user_playback_previous()
}
\references{
\href{https://developer.spotify.com/web-api/skip-users-playback-to-previous-track/}{API documentation}
}
