% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.R
\name{transfer_user_playback}
\alias{transfer_user_playback}
\title{Transfer a User’s Playback
Transfer playback to a new device and determine if it should start playing}
\usage{
transfer_user_playback(device_ids, play = TRUE)
}
\description{
Transfer a User’s Playback
Transfer playback to a new device and determine if it should start playing
}
\examples{
set_tokens()
user_auth()
transfer_user_playback(device_ids = '7c9c590cb73f430fa049c1ccbd5be87334580827')
}
\references{
\href{https://developer.spotify.com/web-api/transfer-a-users-playback/}{API documentation}
}
