% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{validate_country}
\alias{validate_country}
\title{Validate country parameter}
\usage{
validate_country(country)
}
\arguments{
\item{country}{Optional. \cr
An \href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2}{ISO 3166-1 alpha-2 country code} or the string \code{"from_token"}. Provide this parameter if you want to apply \href{https://developer.spotify.com/documentation/general/guides/track-relinking-guide/}{Track Relinking}}
}
\value{
A boolean if the parameter matches the Spotify Web API parameter range.
}
\description{
Validate country parameter
}
\keyword{internal}
