% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{validate_include_meta_info}
\alias{validate_include_meta_info}
\title{Validate include_meta_info parameter}
\usage{
validate_include_meta_info(include_meta_info)
}
\arguments{
\item{include_meta_info}{Optional. Boolean indicating whether to include full result,
with meta information such as \code{"total"}, and \code{"limit"}. Defaults to \code{FALSE}.}
}
\value{
A boolean if the parameter matches the Spotify Web API parameter range.
}
\description{
Validate include_meta_info parameter
}
\keyword{internal}
