% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{validate_limit}
\alias{validate_limit}
\title{Validate limit paramater}
\usage{
validate_limit(limit)
}
\arguments{
\item{limit}{Optional. The maximum number of items to return.
Default to \code{20}. Minimum: 1. Maximum: 50.}
}
\value{
A boolean if the parameter matches the Spotify Web API parameter range.
}
\description{
Validate limit paramater
}
\keyword{internal}
