% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{validate_offset}
\alias{validate_offset}
\title{Validate offset parameter}
\usage{
validate_offset(offset)
}
\arguments{
\item{offset}{Optional. The index of the first item to return.
Defaults to \code{0}, i.e., the first object.
Use with \code{limit} to get the next set of items.}
}
\value{
A boolean if the parameter matches the Spotify Web API parameter range.
}
\description{
Validate offset parameter
}
\keyword{internal}
