% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{validate_position_ms}
\alias{validate_position_ms}
\title{validate position_ms parameter}
\usage{
validate_position_ms(position_ms)
}
\arguments{
\item{position_ms}{Optional. Integer indicating from what position to start playback. Must be a positive number. Passing in a position that is greater than the length of the track will cause the player to start playing the next song.}
}
\value{
A boolean if the parameter matches the Spotify Web API parameter range.
}
\description{
validate position_ms parameter
}
\keyword{internal}
