% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{validate_state}
\alias{validate_state}
\title{validate state paramater}
\usage{
validate_state(state)
}
\arguments{
\item{state}{The state parameter must be exactly one of \code{'track'},
\code{'context'} or  \code{'off'}.}
}
\value{
A boolean if the parameter matches the Spotify Web API parameter range.
}
\description{
validate state paramater
}
\keyword{internal}
