% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{validate_time_range}
\alias{validate_time_range}
\title{validate time_range parameter}
\usage{
validate_time_range(time_range)
}
\value{
A boolean if the parameter matches the Spotify Web API parameter range.
}
\description{
validate time_range parameter
}
\keyword{internal}
