% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{validate_type_artist_or_user}
\alias{validate_type_artist_or_user}
\title{Validate type paramter for 'artists' or 'user'.}
\usage{
validate_type_artist_or_user(artist_or_user)
}
\arguments{
\item{artist_or_user}{"The type parameter must be either 'artist' or 'user'."}
}
\value{
A boolean if the parameter matches the Spotify Web API parameter range.
}
\description{
Validate type paramter for 'artists' or 'user'.
}
\keyword{internal}
