% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{validate_type_artists_or_tracks}
\alias{validate_type_artists_or_tracks}
\title{Validate type paramter for 'artists' or 'tracks'}
\usage{
validate_type_artists_or_tracks(artists_or_tracks)
}
\arguments{
\item{artists_or_tracks}{The type parameter must be either 'artists' or 'tracks'."}
}
\value{
A boolean if the parameter matches the Spotify Web API parameter range.
}
\description{
Validate type paramter for 'artists' or 'tracks'
}
\keyword{internal}
