% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{validate_volume_percent}
\alias{validate_volume_percent}
\title{Validate volume percent parameter}
\usage{
validate_volume_percent(volume_percent)
}
\arguments{
\item{volume_percent}{Required integer value. The volume to set.
Must be a value from 0 to 100 inclusive. Defaults to \code{50}.}
}
\value{
A boolean if the parameter matches the Spotify Web API parameter range.
}
\description{
Validate volume percent parameter
}
\keyword{internal}
