% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{verify_result}
\alias{verify_result}
\title{Verify API Result}
\usage{
verify_result(res)
}
\arguments{
\item{res}{API result ot check}
}
\value{
A meaningful error message if communication with the Spotify Web API was not
successful.
}
\description{
Check API result for error codes.
}
\keyword{internal}
