% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/playlists.R
\name{add_tracks_to_playlist}
\alias{add_tracks_to_playlist}
\title{Add Tracks to a Playlist
Add one or more tracks to a users playlist}
\usage{
add_tracks_to_playlist(user_id, playlist_id, uris)
}
\description{
For more information: https://developer.spotify.com/web-api/add-tracks-to-playlist/
}

