% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playlists.R
\name{add_tracks_to_playlist}
\alias{add_tracks_to_playlist}
\title{Add Tracks to a Playlist
Add one or more tracks to a users playlist}
\usage{
add_tracks_to_playlist(user_id, playlist_id, uris)
}
\arguments{
\item{uris}{}
}
\description{
Add Tracks to a Playlist
Add one or more tracks to a users playlist
}
\examples{
set_tokens()
user_auth()
add_tracks_to_playlist(user_id = 'rweyant', playlist_id = '4Ej1T0FVGBCmwXapPGK2eP', uris = 'spotify:track:4S47FRAZ3VcyZS87FnQtmk')
add_tracks_to_playlist(user_id = 'rweyant',
                       playlist_id = '4Ej1T0FVGBCmwXapPGK2eP',
                       uris = c('spotify:track:4S47FRAZ3VcyZS87FnQtmk', 'spotify:track:3MXA2BkBk0lSuMpRoM7SK2'))
}
\references{
\href{https://developer.spotify.com/web-api/add-tracks-to-playlist/}{API documentation}
}
