% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/library.R
\name{check_saved_tracks}
\alias{check_saved_tracks}
\title{Check User’s Saved Tracks
Check if one or more tracks is already saved in the current Spotify user’s “Your Music” library.}
\usage{
check_saved_tracks(ids, ...)
}
\description{
For more information: https://developer.spotify.com/web-api/get-users-saved-tracks/
}

