% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playlists.R
\name{create_playlist}
\alias{create_playlist}
\title{Create a Playlist
Create a playlist for a Spotify user. (The playlist will be empty until you add tracks.)}
\usage{
create_playlist(user_id, name, ...)
}
\arguments{
\item{user_id}{The user's Spotify user ID.}

\item{name}{Required. The name for the new playlist, for example "Your
Coolest Playlist". This name does not need to be unique; a user may have
several playlists with the same name.}
}
\description{
Create a Playlist
Create a playlist for a Spotify user. (The playlist will be empty until you add tracks.)
}
\examples{
set_tokens()
user_auth()
create_playlist(user_id = 'rweyant', name = 'spotifyr-test')
}
\references{
\href{https://developer.spotify.com/web-api/create-playlist/}{API documentation}
}
