% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/playlists.R
\name{create_playlist}
\alias{create_playlist}
\title{Create a Playlist
Create a playlist for a Spotify user. (The playlist will be empty until you add tracks.)}
\usage{
create_playlist(user_id, name, ...)
}
\description{
For more information: https://developer.spotify.com/web-api/create-playlist/
}

