% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/follow.R
\name{follow}
\alias{follow}
\title{Follow Artists or Users
Add the current user as a follower of one or more artists or other Spotify users.}
\usage{
follow(ids, type = c("artist", "user"), ...)
}
\arguments{
\item{ids}{Optional. A comma-separated list of the artist or the user Spotify IDs.}

\item{type}{Required. The ID type: either artist or user.}
}
\description{
Follow Artists or Users
Add the current user as a follower of one or more artists or other Spotify users.
}
\examples{
set_tokens()
user_auth()
follow('08td7MxkoHQkXnWAYD8d6Q', type='artist')
}
\references{
\href{https://developer.spotify.com/web-api/follow-artists-users/}{API documentation}
}
