% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/follow.R
\name{follow_playlist}
\alias{follow_playlist}
\title{Follow a Playlist
Add the current user as a follower of a playlist}
\usage{
follow_playlist(owner_id, playlist_id, public = TRUE, ...)
}
\arguments{
\item{public}{}
}
\description{
Follow a Playlist
Add the current user as a follower of a playlist
}
\examples{
set_tokens()
user_auth()
follow_playlist(owner_id = 'jmperezperez', playlist_id = '2v3iNvBX8Ay1Gt2uXtUKUT')
}
\references{
\href{https://developer.spotify.com/web-api/follow-playlist/}{API documentation}
}
