% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/follow.R
\name{follow_playlist}
\alias{follow_playlist}
\title{Follow a Playlist
Add the current user as a follower of a playlist}
\usage{
follow_playlist(owner_id, playlist_id, ...)
}
\description{
For more information: https://developer.spotify.com/web-api/follow-playlist/
}

