% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/follow.R
\name{following_playlist}
\alias{following_playlist}
\title{Check if Users Follow a Playlist
Check to see if one or more Spotify users are following a specified playlist.}
\usage{
following_playlist(owner_id, playlist_id, ids, ...)
}
\arguments{
\item{ids}{}
}
\description{
Check if Users Follow a Playlist
Check to see if one or more Spotify users are following a specified playlist.
}
\examples{
set_tokens()
user_auth()
following_playlist(owner_id = 'jmperezperez', playlist_id = '2v3iNvBX8Ay1Gt2uXtUKUT', ids='rweyant')
}
\references{
\href{https://developer.spotify.com/web-api/check-user-following-playlist/}{API documentation}
}
