% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/albums.R
\name{get_album}
\alias{get_album}
\title{Get an Album
Get Spotify catalog information for a single album}
\usage{
get_album(id, ...)
}
\arguments{
\item{id}{Required. The Spotify ID for the album}
}
\description{
Get an Album
Get Spotify catalog information for a single album
}
\examples{
set_tokens()
get_album('0sNOF9WDwhWunNAHPD3Baj')
get_album('0sNOF9WDwhWunNAHPD3Baj', market='IS')
}
\references{
\href{https://developer.spotify.com/web-api/get-album/}{API documentation}
}
