% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_albums.R
\name{get_albums}
\alias{get_albums}
\title{Get Albums}
\usage{
get_albums(artist_uri, access_token = get_spotify_access_token())
}
\arguments{
\item{artist_uri}{String identifier for an artist on Spotify. Can be found within the Spotify app or with spotifyr::get_artists()}

\item{access_token}{Spotify Web API token. Defaults to spotifyr::get_spotify_access_token()}
}
\description{
This function returns an artist's discography on Spotify
}
\examples{
\dontrun{
artists <- get_artists('radiohead')
albums <- get_albums(artists$artist_uri[1])
}
}
\keyword{albums}
