% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/albums.R
\name{get_albums}
\alias{get_albums}
\title{Get Several Albums
Get Spotify catalog information for multiple albums identified by their Spotify IDs.}
\usage{
get_albums(ids, ...)
}
\arguments{
\item{ids}{Required. A comma-separated list of the Spotify IDs for the albums. Maximum: 20 IDs.}
}
\description{
Get Several Albums
Get Spotify catalog information for multiple albums identified by their Spotify IDs.
}
\examples{
set_tokens()
get_albums(ids = c('41MnTivkwTO3UUJ8DrqEJJ', '6JWc4iAiJ9FjyK0B59ABb4', '6UXCm6bOO4gFlDQZV5yL37'))
get_albums(ids = c('41MnTivkwTO3UUJ8DrqEJJ', '6JWc4iAiJ9FjyK0B59ABb4', '6UXCm6bOO4gFlDQZV5yL37'), market='US')
}
\references{
\href{https://developer.spotify.com/web-api/get-several-album/}{API documentation}
}
