% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/albums.R
\name{get_albums}
\alias{get_albums}
\title{Get Several Albums
Get Spotify catalog information for multiple albums identified by their Spotify IDs.}
\usage{
get_albums(ids, ...)
}
\arguments{
\item{ids}{Required. A comma-separated list of the Spotify IDs for the albums. Maximum: 20 IDs.

For more information: https://developer.spotify.com/web-api/get-several-albums/}
}
\description{
Get Several Albums
Get Spotify catalog information for multiple albums identified by their Spotify IDs.
}

