% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/artists.R
\name{get_artist}
\alias{get_artist}
\title{Get an Artist
Get Spotify catalog information for a single artist identified by their unique Spotify ID.}
\usage{
get_artist(id)
}
\arguments{
\item{id}{Required. The Spotify ID for the album

For more information: https://developer.spotify.com/web-api/get-artist/}
}
\description{
Get an Artist
Get Spotify catalog information for a single artist identified by their unique Spotify ID.
}

