% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artists.R
\name{get_artist}
\alias{get_artist}
\title{Get an Artist}
\usage{
get_artist(id, ...)
}
\arguments{
\item{id}{Required. The Spotify ID for the album}
}
\value{
Spotify Artist object
}
\description{
Get Spotify catalog information for a single artist identified by their unique Spotify ID.
}
\examples{
set_tokens()
get_artist('42ZTPxJukaCRQuMjmxVJo4')
}
\references{
\href{https://developer.spotify.com/web-api/get-artist/}{API documentation}
}
