% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_artist_albums.R
\name{get_artist_albums}
\alias{get_artist_albums}
\title{Get Artist Albums}
\usage{
get_artist_albums(artist_name = NULL, artist_uri = NULL,
  use_artist_uri = FALSE, return_closest_artist = TRUE, message = FALSE,
  studio_albums_only = TRUE, access_token = get_spotify_access_token())
}
\arguments{
\item{artist_name}{String of artist name}

\item{artist_uri}{String of Spotify artist URI. Will only be applied if \code{use_arist_uri} is set to \code{TRUE}. This is useful for pulling related artists in bulk and allows for more accurate matching since Spotify URIs are unique.}

\item{use_artist_uri}{Boolean determining whether to search by Spotify URI instead of an artist name. If \code{TRUE}, you must also enter an \code{artist_uri}. Defaults to \code{FALSE}.}

\item{studio_albums_only}{Logical for whether to remove album types "single" and "compilation" and albums with mulitple artists. Defaults to \code{TRUE}}

\item{access_token}{Spotify Web API token. Defaults to spotifyr::get_spotify_access_token()}
}
\description{
This function returns an artist's discography on Spotify
}
\examples{
\dontrun{
albums <- get_artist_albums('radiohead')
}
}
\keyword{albums}
