% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artists.R
\name{get_artist_albums}
\alias{get_artist_albums}
\title{Get Artist Albums}
\usage{
get_artist_albums(artist = NULL, album_types = "album",
  return_closest_artist = TRUE, access_token = get_spotify_access_token(),
  parallelize = FALSE, future_plan = "multiprocess")
}
\arguments{
\item{artist}{String of artist name or Spotify artist URI}

\item{album_types}{Character vector of album types to include. Valid values are "album", "single", "appears_on", and "compilation". Defaults to "album".}

\item{access_token}{Spotify Web API token. Defaults to spotifyr::get_spotify_access_token()}

\item{parallelize}{Boolean determining to run in parallel or not. Defaults to \code{FALSE}.}

\item{future_plan}{String determining how `future()`s are resolved when `parallelize == TRUE`. Defaults to \code{multiprocess}.}
}
\description{
This function returns an artist's discography on Spotify
}
\examples{
\dontrun{
albums <- get_artist_albums('radiohead')
}
}
\keyword{albums}
