% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/artists.R
\name{get_artist_albums}
\alias{get_artist_albums}
\title{Get an Artist’s Albums
Get Spotify catalog information about an artist’s albums.
Optional parameters can be specified in the query string to
filter and sort the response.}
\usage{
get_artist_albums(id, ...)
}
\arguments{
\item{id}{Required. The Spotify ID for the artist

For more information: https://developer.spotify.com/web-api/get-albums-tracks/}
}
\description{
Get an Artist’s Albums
Get Spotify catalog information about an artist’s albums.
Optional parameters can be specified in the query string to
filter and sort the response.
}

