% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artists.R
\name{get_artist_audio_features}
\alias{get_artist_audio_features}
\title{Get features and popularity for an artist's entire discography on Spotify}
\usage{
get_artist_audio_features(artist = NULL, album_types = "album",
  return_closest_artist = TRUE,
  access_token = get_spotify_access_token(), parallelize = FALSE,
  future_plan = "multiprocess")
}
\arguments{
\item{artist}{String of artist name or Spotify artist URI}

\item{album_types}{Character vector of album types to include. Valid values are "album", "single", "appears_on", and "compilation". Defaults to "album".}

\item{return_closest_artist}{Boolean for selecting the artist result with the closest match on Spotify's Search endpoint. Defaults to \code{TRUE}.}

\item{access_token}{Spotify Web API token. Defaults to spotifyr::get_spotify_access_token()}

\item{parallelize}{Boolean determining to run in parallel or not. Defaults to \code{FALSE}.}

\item{future_plan}{String determining how `future()`s are resolved when `parallelize == TRUE`. Defaults to \code{multiprocess}.}
}
\description{
This function returns the popularity and audio features for every song and album for a given artist on Spotify
}
\examples{
\dontrun{
radiohead_features <- get_artist_audio_features(artist_name = 'radiohead')
}
}
\keyword{audio}
\keyword{discography}
\keyword{features}
\keyword{track}
