% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_artist_audio_features.R
\name{get_artist_audio_features}
\alias{get_artist_audio_features}
\title{Get features and popularity for an artist's entire discography on Spotify}
\usage{
get_artist_audio_features(artist_name,
  access_token = get_spotify_access_token(), return_closest_artist = TRUE,
  message = FALSE)
}
\arguments{
\item{artist_name}{String of artist name}

\item{access_token}{Spotify Web API token. Defaults to spotifyr::get_spotify_access_token()}

\item{return_closest_artist}{Boolean for using string distance to select the artist result with the closest match to the given string. Defaults to \code{TRUE}.}

\item{message}{Boolean for printing the name of artist matched when using \code{return_closest_artist = TRUE}. Defaults to \code{FALSE}.}
}
\description{
This function returns the popularity and audio features for every song and album for a given artist on Spotify
}
\examples{
\dontrun{
radiohead_features <- get_artist_audio_features('radiohead')
}
}
\keyword{audio}
\keyword{discography}
\keyword{features}
\keyword{track}
