% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_artist_audio_features.R
\name{get_artist_audio_features}
\alias{get_artist_audio_features}
\title{Get features and popularity for an artist's entire discography on Spotify}
\usage{
get_artist_audio_features(artist_name,
  access_token = get_spotify_access_token())
}
\arguments{
\item{artist_name}{String of artist name}

\item{access_token}{Spotify Web API token. Defaults to spotifyr::get_spotify_access_token()}
}
\description{
This function returns the popularity and audio features for every song and album for a given artist on Spotify
}
\examples{
radiohead_features <- get_artist_audio_features('radiohead')
}
\keyword{audio}
\keyword{discography}
\keyword{features}
\keyword{track}
