% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/artists.R
\name{get_artist_toptracks}
\alias{get_artist_toptracks}
\title{Get an Artist’s Top Tracks
Get Spotify catalog information about an artist’s top tracks by country.
filter and sort the response.}
\usage{
get_artist_toptracks(id, country)
}
\arguments{
\item{id}{Required. The Spotify ID for the artist

For more information: https://developer.spotify.com/web-api/get-artists-top-tracks/}
}
\description{
Get an Artist’s Top Tracks
Get Spotify catalog information about an artist’s top tracks by country.
filter and sort the response.
}

