% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artists.R
\name{get_artists}
\alias{get_artists}
\title{Get Several Artists
Get Spotify catalog information for several artists based on their Spotify IDs.}
\usage{
get_artists(ids)
}
\arguments{
\item{ids}{Required. A comma-separated list of the Spotify IDs for the artists. Maximum: 50 IDs.}
}
\description{
Get Several Artists
Get Spotify catalog information for several artists based on their Spotify IDs.
}
\examples{
get_artists(c('0oSGxfWSnnOXhD2fKuz2Gy','3dBVyJ7JuOMt4GE9607Qin'))
}
\references{
\href{https://developer.spotify.com/web-api/get-several-artists/}{API documentation}
}
