% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/browse.R
\name{get_categories}
\alias{get_categories}
\title{Get a List of Categories
Get a list of categories used to tag items in Spotify (on, for example, the Spotify player’s “Browse” tab).}
\usage{
get_categories(...)
}
\description{
For more information: https://developer.spotify.com/web-api/get-list-categories/
}

