% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{get_category}
\alias{get_category}
\title{Get a Category
Get a single category used to tag items in Spotify (on, for example, the Spotify player’s “Browse” tab).}
\usage{
get_category(category_id, ...)
}
\arguments{
\item{category_id}{The Spotify category ID for the category.}
}
\description{
For more information: https://developer.spotify.com/web-api/get-category/
}
\examples{
get_category('party')
}
