% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{get_category_playlists}
\alias{get_category_playlists}
\title{Get a Category’s Playlists
Get a list of Spotify playlists tagged with a particular category.}
\usage{
get_category_playlists(category_id, ...)
}
\arguments{
\item{category_id}{The Spotify category ID for the category.}
}
\description{
Get a Category’s Playlists
Get a list of Spotify playlists tagged with a particular category.
}
\examples{
get_category_playlists('party')
}
\references{
\href{https://developer.spotify.com/web-api/get-categorys-playlists/}{API documentation}
}
