% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/browse.R
\name{get_category_playlists}
\alias{get_category_playlists}
\title{Get a Category’s Playlists
Get a list of Spotify playlists tagged with a particular category.}
\usage{
get_category_playlists(category_id, ...)
}
\description{
For more information: https://developer.spotify.com/web-api/get-categorys-playlists/
}

