% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{get_current_user_profile}
\alias{get_current_user_profile}
\title{Get Current User’s Profile
Get detailed profile information about the current user (including the current user’s username).}
\usage{
get_current_user_profile()
}
\description{
Get Current User’s Profile
Get detailed profile information about the current user (including the current user’s username).
}
\examples{
set_tokens()
user_auth()
get_current_user_profile()
}
\references{
\href{https://developer.spotify.com/web-api/get-current-users-profile/}{API documentation}
}
