% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playlists.R
\name{get_playlist}
\alias{get_playlist}
\title{Get a Playlist
Get a playlist owned by a Spotify user.}
\usage{
get_playlist(user_id, playlist_id, ...)
}
\arguments{
\item{playlist_id}{}
}
\description{
Get a Playlist
Get a playlist owned by a Spotify user.
}
\examples{
set_tokens()
get_playlist(user_id = 'spotify', playlist_id = '59ZbFPES4DQwEjBpWHzrtC')
}
\references{
\href{https://developer.spotify.com/web-api/get-playlist/}{API documentation}
}
