% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playlists.R
\name{get_playlist_tracks}
\alias{get_playlist_tracks}
\title{Get a Playlist’s Tracks
Get full details of the tracks of a playlist owned by a Spotify user.}
\usage{
get_playlist_tracks(user_id, playlist_id, ...)
}
\arguments{
\item{user_id}{The user's Spotify user ID.}

\item{playlist_id}{The Spotify ID for the playlist.}
}
\description{
Get a Playlist’s Tracks
Get full details of the tracks of a playlist owned by a Spotify user.
}
\examples{
set_tokens()
get_playlist_tracks(user_id = 'spotify', playlist_id = '59ZbFPES4DQwEjBpWHzrtC')
}
\references{
\href{https://developer.spotify.com/web-api/get-playlists-tracks/}{API documentation}
}
