% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Playlists.R
\name{get_playlists}
\alias{get_playlists}
\alias{get_playlist_tracks}
\alias{add_playlist}
\alias{add_tracks}
\title{Manage Playlists}
\usage{
get_playlists(user, private = F)

get_playlist_tracks(playlist)

add_playlist(name)

add_tracks(ids, playlist)
}
\arguments{
\item{user}{Username of owner}

\item{private}{Include private playlists?}

\item{playlist}{ID of playlist}

\item{name}{Name of playlist}

\item{ids}{Vector of track IDs}
}
\description{
Manage Playlists
}
\section{Functions}{
\itemize{
\item \code{get_playlist_tracks}: 

\item \code{add_playlist}: 

\item \code{add_tracks}: 
}}

