% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/library.R
\name{get_saved_tracks}
\alias{get_saved_tracks}
\title{Get a User’s Saved Tracks
Get a list of the songs saved in the current Spotify user’s “Your Music” library.}
\usage{
get_saved_tracks(access_token = NULL, ...)
}
\description{
For more information: https://developer.spotify.com/web-api/get-users-saved-tracks/
}

