% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks.R
\name{get_track}
\alias{get_track}
\title{Get a Track
Get Spotify catalog information for a single track identified by its unique Spotify ID.}
\usage{
get_track(id, ...)
}
\arguments{
\item{id}{}
}
\description{
Get a Track
Get Spotify catalog information for a single track identified by its unique Spotify ID.
}
\examples{
set_tokens()
get_track(id='3n3Ppam7vgaVa1iaRUc9Lp')
}
\references{
\href{https://developer.spotify.com/web-api/get-track/}{API documentation}
}
