% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks.R
\name{get_tracks}
\alias{get_tracks}
\title{Get Several Tracks
Get Spotify catalog information for multiple tracks based on their Spotify IDs.}
\usage{
get_tracks(ids, ...)
}
\arguments{
\item{ids}{A comma-separated list of the Spotify IDs for the tracks. Maximum: 50 IDs.}
}
\description{
Get Several Tracks
Get Spotify catalog information for multiple tracks based on their Spotify IDs.
}
\examples{
set_tokens()
get_tracks(ids='3n3Ppam7vgaVa1iaRUc9Lp')
get_tracks(ids=c('3n3Ppam7vgaVa1iaRUc9Lp', '3twNvmDtFQtAd5gMKedhLD'))
}
\references{
\href{https://developer.spotify.com/web-api/get-several-tracks/}{API documentation}
}
