% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{get_user_audio_features}
\alias{get_user_audio_features}
\title{Get audio feature information for a users' playlists}
\usage{
get_user_audio_features(username = NULL,
  authorization = get_spotify_access_token())
}
\arguments{
\item{username}{Required. String of Spotify username.}

\item{authorization}{Required. A valid access token from the Spotify Accounts service. See the \href{https://developer.spotify.com/documentation/general/guides/authorization-guide/}{Web API authorization Guide} for more details. Defaults to \code{spotifyr::get_spotify_access_token()}}
}
\value{
Returns a data frame of results containing track audio features data. See the \href{https://developer.spotify.com/documentation/web-api/reference/tracks/get-several-audio-features/}{Spotify Web API documentation} for more information.
}
\description{
Get audio feature information for a users' playlists
}
