% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_playlist_count.R
\name{get_user_playlist_count}
\alias{get_user_playlist_count}
\title{Get count of Spotify playlists for a given user}
\usage{
get_user_playlist_count(username, access_token = get_spotify_access_token(),
  echo = F)
}
\arguments{
\item{username}{String of Spotify username. Can be found within the Spotify App}

\item{access_token}{Spotify Web API token. Defaults to spotifyr::get_spotify_access_token()}
}
\description{
Helper function for spotifyr::get_user_playlists()
}
\examples{
obama_playlist_count <- get_user_playlist_count('barackobama')
}
\keyword{username}
