% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playlists.R
\name{get_user_playlists}
\alias{get_user_playlists}
\title{Get user playlists}
\usage{
get_user_playlists(username, access_token = get_spotify_access_token(),
  parallelize = FALSE, future_plan = "multiprocess")
}
\arguments{
\item{username}{String of Spotify username. Can be found within the Spotify App}

\item{access_token}{Spotify Web API token. Defaults to \code{spotifyr::get_spotify_access_token()}.}

\item{parallelize}{Boolean determining to run in parallel or not. Defaults to \code{FALSE}.}

\item{future_plan}{String determining how `future()`s are resolved when `parallelize == TRUE`. Defaults to \code{multiprocess}.}
}
\description{
This function returns a dataframe of playlists for a given Spotify username
}
\examples{
\dontrun{
get_user_playlists('barackobama')
}
}
\keyword{username}
