% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/profiles.R
\name{get_user_profile}
\alias{get_user_profile}
\title{Get a User’s Profile
Get public profile information about a Spotify user}
\usage{
get_user_profile(user_id)
}
\description{
For more information: https://developer.spotify.com/web-api/get-users-profile/
}

