% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{get_user_profile}
\alias{get_user_profile}
\title{Get a User’s Profile
Get public profile information about a Spotify user}
\usage{
get_user_profile(user_id)
}
\arguments{
\item{user_id}{}
}
\description{
Get a User’s Profile
Get public profile information about a Spotify user
}
\examples{
set_tokens()
get_user_profile(user_id = 'rweyant')
}
\references{
\href{https://developer.spotify.com/web-api/get-users-profile/}{API documentation}
}
