% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{pitch_class_lookup}
\alias{pitch_class_lookup}
\title{Pitch class notation lookup}
\format{An object of class \code{character} of length 12.}
\usage{
pitch_class_lookup
}
\description{
Pitch class notation lookup
}
\keyword{datasets}
