% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{remove_track}
\alias{remove_track}
\title{Remove User’s Saved Tracks
Remove one or more tracks from the current user’s “Your Music” library}
\usage{
remove_track(ids, ...)
}
\arguments{
\item{ids}{}
}
\description{
Remove User’s Saved Tracks
Remove one or more tracks from the current user’s “Your Music” library
}
\examples{
set_tokens()
user_auth()
remove_track(ids=c('4iV5W9uYEdYUVa79Axb7Rh'))
}
\references{
\href{https://developer.spotify.com/web-api/remove-tracks-user/}{API documentation}
}
