% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playlists.R
\name{remove_tracks_from_playlist}
\alias{remove_tracks_from_playlist}
\title{Remove Tracks from a Playlist
Remove one or more tracks from a user’s playlist.}
\usage{
remove_tracks_from_playlist(user_id, playlist_id, uris)
}
\arguments{
\item{uris}{}
}
\description{
Remove Tracks from a Playlist
Remove one or more tracks from a user’s playlist.
}
\examples{
set_tokens()
user_auth()
remove_tracks_from_playlist(user_id = 'rweyant', playlist_id = '4Ej1T0FVGBCmwXapPGK2eP', uris = 'spotify:track:4S47FRAZ3VcyZS87FnQtmk')
remove_tracks_from_playlist(user_id = 'rweyant',
                            playlist_id = '4Ej1T0FVGBCmwXapPGK2eP',
                            uris = c('spotify:track:4S47FRAZ3VcyZS87FnQtmk', 'spotify:track:3MXA2BkBk0lSuMpRoM7SK2'))
}
\references{
\href{https://developer.spotify.com/web-api/remove-tracks-playlist/}{API documentation}
}
