% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/library.R
\name{save_track}
\alias{save_track}
\title{Save Tracks for User
Save one or more tracks to the current user’s “Your Music” library}
\usage{
save_track(ids, ...)
}
\description{
For more information: https://developer.spotify.com/web-api/save-tracks-user/
}

