% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{scopes}
\alias{scopes}
\title{Valid Authorization Scopes}
\format{
An object of class \code{character} of length 19.
}
\usage{
scopes
}
\value{
A character vector of valid authorization scopes for the Spotify Web API.
See \href{https://developer.spotify.com/documentation/general/guides/scopes/}{Spotify Web API Authorization Scopes}
}
\description{
A vector of valid scopes for \code{\link{get_authorization_code}}
}
\examples{
scopes()
}
\concept{authorization functions}
\keyword{datasets}
