% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scopes}
\alias{scopes}
\title{Valid Authorization Scopes}
\usage{
scopes(exclude_SOA = TRUE)
}
\arguments{
\item{exclude_SOA}{Boolean indicating whether to exclude 'Spotify Open Access' (SOA) scopes, which are only available for approved partners. Defaults to \code{TRUE}.}
}
\value{
A character vector of valid authorization scopes for the Spotify Web API.
See \href{https://developer.spotify.com/documentation/general/guides/authorization/scopes/}{Spotify Web API Authorization Scopes}
}
\description{
A vector of valid scopes for \code{\link{get_spotify_authorization_code}}.
}
\examples{
scopes()
}
\concept{authorization functions}
