% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search}
\alias{search}
\title{Get Spotify catalog information about artists, albums,
tracks or playlists that match a keyword string.}
\usage{
search(q, type = c("artist", "album", "playlist", "track"), ...)
}
\arguments{
\item{q}{Required. The search query's keywords
(and optional field filters and operators)}

\item{type}{A comma-separated list of item types
to search across. Valid types are: album, artist,
playlist, and track}
}
\description{
Get Spotify catalog information about artists, albums,
tracks or playlists that match a keyword string.
}
\examples{
set_tokens()
search(q = 'heavy metal', type = 'track')
}
\references{
\href{https://developer.spotify.com/web-api/search-item/}{API documentation}
}
