% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Search.R
\name{search_spotify}
\alias{search_spotify}
\title{Search Spotify}
\usage{
search_spotify(query, type = "track", limit = 1)
}
\arguments{
\item{query}{Search query. Supports track:, artist:, etc.}

\item{type}{Result type--one of track, artist, album, playlist}

\item{limit}{Number of results to return (max 50)}
}
\description{
Search Spotify
}
