% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{set_credentials}
\alias{set_credentials}
\title{Set credentials to be accessed by all functions}
\usage{
set_credentials(client_id, client_secret, client_redirect_uri)
}
\arguments{
\item{client_id}{the client ID given from Spotify}

\item{client_secret}{the client secret ID given from Spotify}
}
\description{
Set credentials to be accessed by all functions
}
