% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotifyr.R
\docType{package}
\name{spotifyr}
\alias{spotifyr}
\title{\code{spotifyr} package}
\description{
A Quick and Easy Wrapper for Pulling Track Audio Features from Spotify's Web API in Bulk
}
\details{
See the README on
\href{https://github.com/charlie86/spotifyr#readme}{GitHub}
}
\section{artist functions}{

Retrieve information related to artists.
\code{\link{artist}}: Get the data of a single artist.\cr
\code{\link{get_artists}}: Get the data of multiple artist.\cr
\code{\link{get_related_artists}}: Get that of related artist to an original artist.\cr
\code{\link{get_artist_albums}}: Get artists who appear on an album.\cr
}

\section{track functions}{

Retrieve information related to individual song recordings (songs, concerts).
\code{\link{get_track}} \cr
\code{\link{get_tracks}} \cr
\code{\link{get_playlist_tracks}} \cr
}

\section{album functions}{

Retrieve information about albums.
See also \code{\link{get_album_tracks}}.
}

\section{playlist functions}{

Work with playlists.\cr
\code{\link{add_tracks_to_playlist}} \cr
\code{\link{change_playlist_details}} \cr
\code{\link{get_user_playlist}} \cr
\code{\link{get_user_playlists}} \cr
}

\section{player functions}{

Interact with the user's devices and players.
}

\section{personalization functions}{

\code{\link{get_my_top_artists_or_tracks}}
}

\section{musicology functions}{

Functions related to the contents of the music.\cr
\code{\link{get_track_audio_analysis}} \cr
\code{\link{get_artist_audio_features}} \cr
\code{\link{get_playlist_audio_features}} \cr
\code{\link{get_user_audio_features}}\cr
\code{\link{get_genre_artists}} \cr
}

\section{lyrics functions}{

Functions related to the lyrics of the music, such as \cr
\code{\link{get_album_data}} \cr
\code{\link{get_discography}}.
\code{\link{get_playlist_audio_features}}.
}

\section{search functions}{

Search for an artist, song or other keyword.\cr
The main function is \code{\link{search_spotify}}.
}

\section{authentication functions}{

Helper functions to provide authentication function for requests.
Never reveal in public documents, markdown files the returned values of these
functions: \cr
\code{\link{get_spotify_authorization_code}} \cr
\code{\link{get_spotify_access_token}}.
}

