% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotifyr.R
\docType{package}
\name{spotifyr}
\alias{spotifyr}
\title{\code{spotifyr} package}
\description{
A Quick and Easy Wrapper for Pulling Track Audio Features from Spotify's Web API in Bulk
}
\details{
See the README on
\href{https://github.com/charlie86/spotifyr#readme}{GitHub}
}
\section{artist functions}{

Retrieve information related to artists.
\code{artist}: Get the data of a single artist.\cr
\code{get_artists}: Get the data of multiple artist.\cr
\code{get_related_artists}: Get that of related artist to an original artist.\cr
\code{get_artist_albums}: Get artists who appear on an album.\cr
}

\section{track functions}{

Retrieve information related to individual song recordings (songs, concerts).
}

\section{album functions}{

Retrieve information about albums.
See also \code{\link{get_album_tracks}}.
}

\section{playlist functions}{

Work with playlists.
}

\section{player functions}{

Interact with the user's devices and players.
}

\section{personalization functions}{

\code{\link{get_my_top_artists_or_tracks}}
}

\section{musicology functions}{

Functions related to the contents of the music.
\code{get_artist_audio_features} \cr
\code{get_playlist_audio_features} \cr
\code{get_user_audio_features} \cr
\code{get_genre_artists} \cr
}

\section{lyrics functions}{

Functions related to the lyrics of the music, such as
\code{get_album_data} and \code{get_discography}.
\code{\link{get_playlist_audio_features}}.
}

\section{search functions}{

Search for an artist, song or other keyword.\cr
The main function is \code{search_spotify}.
}

\section{authentication functions}{

Helper functions to provide authentication function for requests.
Never reveal in public documents, markdown files the returned values of these
functions:
\code{get_spotify_authorization_code}
\code{get_spotify_access_token}.
}

