% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotifyr.R
\docType{package}
\name{spotifyr}
\alias{spotifyr}
\alias{spotifyr-package}
\title{\code{spotifyr} package}
\description{
A Quick and Easy Wrapper for Pulling Track Audio Features from Spotify's Web API in Bulk
}
\details{
See the README on
\href{https://github.com/charlie86/spotifyr#readme}{GitHub}
}
