% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/follow.R
\name{unfollow}
\alias{unfollow}
\title{Unfollow Artists or Users
Remove the current user as a follower of one or more artists or other Spotify users}
\usage{
unfollow(ids, type = c("artist", "user"), ...)
}
\description{
For more information: https://developer.spotify.com/web-api/unfollow-artists-users/
}

