% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/follow.R
\name{unfollow_playlist}
\alias{unfollow_playlist}
\title{Unfollow a Playlist
Remove the current user as a follower of a playlist.}
\usage{
unfollow_playlist(owner_id, playlist_id, ...)
}
\arguments{
\item{playlist_id}{}
}
\description{
Unfollow a Playlist
Remove the current user as a follower of a playlist.
}
\examples{
set_tokens()
user_auth()
unfollow_playlist(owner_id = 'jmperezperez', playlist_id = '2v3iNvBX8Ay1Gt2uXtUKUT')
}
\references{
\href{https://developer.spotify.com/web-api/unfollow-playlist/}{API documentation}
}
