% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/follow.R
\name{unfollow_playlist}
\alias{unfollow_playlist}
\title{Unfollow a Playlist
Remove the current user as a follower of a playlist.}
\usage{
unfollow_playlist(owner_id, playlist_id, ...)
}
\description{
For more information: https://developer.spotify.com/web-api/unfollow-playlist/
}

